+++ src/cartography/calculations/nomenclature_title.py
@@ -4,9 +4,9 @@ from typing import Any
 import errors
 from misc import from_tuple
 
-from business.math_actions import coordinate_actions
-from business.models import CoordinatePair, Nomenclature
-from business.types import NomenclatureTitleFormatter
+from cartography.math_actions import coordinate_actions
+from cartography.models import CoordinatePair, Nomenclature
+from cartography.types import NomenclatureTitleFormatter
 
 
 def get_1m_nomenclature(nomenclature_title: str) -> Nomenclature:
