+++ src/cartography/chain/interfaces.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 import abc
 
-from business.enums import Scale
-from business.models import CoordinatePair, Nomenclature
-from business.types import NomenclatureTitleDictType, NomenclatureTitleFormatter
+from cartography.enums import Scale
+from cartography.models import CoordinatePair, Nomenclature
+from cartography.types import NomenclatureTitleDictType, NomenclatureTitleFormatter
 
 
 class ICoordinateChainLink(abc.ABC):
