+++ src/cartography/chain/nomenclature_title_chain.py
@@ -1,8 +1,8 @@
-from business import constants
-from business.calculations.nomenclature_title import get_1m_nomenclature, get_nomenclature_by_parts
-from business.enums import Scale
-from business.models import Nomenclature
-from business.types import NomenclatureTitleDictType
+from cartography import constants
+from cartography.calculations.nomenclature_title import get_1m_nomenclature, get_nomenclature_by_parts
+from cartography.enums import Scale
+from cartography.models import Nomenclature
+from cartography.types import NomenclatureTitleDictType
 
 from .interfaces import INomenclatureTitleChainLink
