+++ src/cartography/image_drawer/__init__.py
