+++ src/cartography/image_drawer/labels_drawer.py
@@ -4,9 +4,9 @@ from container import ImageContainer
 from dependency_injector.wiring import Provide, inject
 from PIL import Image, ImageDraw, ImageFont
 
-from business.math_actions import coordinate_actions
-from business.models import Coordinate
-from business.types import ImageColorType
+from cartography.math_actions import coordinate_actions
+from cartography.models import Coordinate
+from cartography.types import ImageColorType
 
 
 @inject
