+++ src/cartography/math_actions/coordinate_actions.py
@@ -1,6 +1,6 @@
 from decimal import Decimal
 
-from business.models import Coordinate
+from cartography.models import Coordinate
 
 
 def _transform_to_coordinate(seconds: Decimal) -> Coordinate:
