+++ src/cartography/scale_resolver/coordinate_chain_link_resolver.py
@@ -1,6 +1,6 @@
-from business.chain import coordinate_chain
-from business.chain.interfaces import ICoordinateChainLink
-from business.enums import Scale
+from cartography.chain import coordinate_chain
+from cartography.chain.interfaces import ICoordinateChainLink
+from cartography.enums import Scale
 
 chain_list: list[type[ICoordinateChainLink]] = [
     coordinate_chain.ChainLink1M,
