+++ src/cartography/scale_resolver/coordinate_resolver.py
@@ -1,5 +1,5 @@
-from business.chain import coordinate_chain
-from business.chain.interfaces import ICoordinateChainLink
+from cartography.chain import coordinate_chain
+from cartography.chain.interfaces import ICoordinateChainLink
 
 table: dict[int, type[ICoordinateChainLink]] = {
     1: coordinate_chain.ChainLink1M,
