+++ src/cartography/scale_resolver/nomenclature_title_chain_link_resolver.py
@@ -1,6 +1,6 @@
-from business.chain import nomenclature_title_chain
-from business.chain.interfaces import INomenclatureTitleChainLink
-from business.enums import Scale
+from cartography.chain import nomenclature_title_chain
+from cartography.chain.interfaces import INomenclatureTitleChainLink
+from cartography.enums import Scale
 
 chain_list: list[type[INomenclatureTitleChainLink]] = [
     nomenclature_title_chain.ChainLink1M,
