+++ src/tg_bot/handlers/geography_coordinates.py
@@ -5,8 +5,8 @@ from aiogram import Router, flags
 from aiogram.filters import Command
 from aiogram.fsm.context import FSMContext
 from aiogram.types import BufferedInputFile, InputMediaPhoto, Message
-from business.facades.nomenclature_facade import NomenclatureFacade
-from business.models import CoordinatePair
+from cartography.facades.nomenclature_facade import NomenclatureFacade
+from cartography.models import CoordinatePair
 from misc import decorators
 from tg_bot.states import ByCoordinatesImages
