+++ tests/test_coordinate_pair_actions.py
@@ -1,6 +1,6 @@
 import pytest
-from business.math_actions.coordinate_pair_actions import get_delta
-from business.models import CoordinatePair
+from cartography.math_actions.coordinate_pair_actions import get_delta
+from cartography.models import CoordinatePair
 from misc import from_tuple
