+++ tests/test_helper.py
@@ -1,4 +1,4 @@
-from business.models import CoordinatePair
+from cartography.models import CoordinatePair
 from misc import from_tuple
